Decouple Config from reqwest
