+++ kube/src/config/incluster_config.rs
@@ -1,9 +1,9 @@
 use std::env;
 
-use crate::{Error, Result};
-use reqwest::Certificate;
-
-use crate::{config::utils, error::ConfigError};
+use crate::{
+    config::{utils, Der},
+    Result,
+};
 
 pub const SERVICE_HOSTENV: &str = "KUBERNETES_SERVICE_HOST";
 pub const SERVICE_PORTENV: &str = "KUBERNETES_SERVICE_PORT";
@@ -32,17 +32,18 @@ pub fn load_token() -> Result<String> {
 }
 
 /// Returns certification from specified path in cluster.
-pub fn load_cert() -> Result<Vec<Certificate>> {
+pub fn load_cert() -> Result<Vec<Der>> {
     let ca = utils::data_or_file_with_base64(&None, &Some(SERVICE_CERTFILE))?;
-    let pems = pem::parse_many(ca);
-
-    pems.into_iter()
-        .map(|pem| {
-            Certificate::from_pem(&pem::encode(&pem).into_bytes())
-                .map_err(ConfigError::LoadCert)
-                .map_err(Error::from)
+    Ok(pem::parse_many(&ca)
+        .into_iter()
+        .filter_map(|p| {
+            if p.tag == "CERTIFICATE" {
+                Some(Der(p.contents))
+            } else {
+                None
+            }
         })
-        .collect::<Result<Vec<_>>>()
+        .collect::<Vec<_>>())
 }
 
 /// Returns the default namespace from specified path in cluster.
