+++ kube/src/config/mod.rs
@@ -10,15 +10,33 @@ mod file_loader;
 mod incluster_config;
 mod utils;
 
-use crate::{error::ConfigError, Result};
+use crate::{error::ConfigError, Error, Result};
+use file_loader::ConfigLoader;
 pub use file_loader::KubeConfigOptions;
-use file_loader::{ConfigLoader, Der};
 
 use chrono::{DateTime, Utc};
-use reqwest::header::{self, HeaderMap};
+use reqwest::{
+    header::{self, HeaderMap},
+    Certificate,
+};
 use tokio::sync::Mutex;
 
 use std::{sync::Arc, time::Duration};
+
+/// Regardless of tls type, a Certificate Der is always a byte array
+#[derive(Debug, Clone)]
+pub struct Der(pub Vec<u8>);
+
+use std::convert::TryFrom;
+impl TryFrom<Der> for Certificate {
+    type Error = Error;
+
+    fn try_from(val: Der) -> Result<Certificate> {
+        Certificate::from_der(&val.0)
+            .map_err(ConfigError::LoadCert)
+            .map_err(Error::from)
+    }
+}
 
 #[derive(Debug, Clone)]
 #[allow(clippy::large_enum_variant)]
@@ -70,7 +88,7 @@ pub struct Config {
     /// The configured default namespace
     pub default_ns: String,
     /// The configured root certificate
-    pub root_cert: Option<Vec<reqwest::Certificate>>,
+    pub root_cert: Option<Vec<Der>>,
     /// Default headers to be used to communicate with the Kubernetes API
     pub headers: HeaderMap,
     /// Timeout for calls to the Kubernetes API.
@@ -205,16 +223,10 @@ impl Config {
         let mut identity = None;
 
         if let Some(ca_bundle) = loader.ca_bundle()? {
-            use std::convert::TryInto;
             for ca in &ca_bundle {
                 accept_invalid_certs = hacky_cert_lifetime_for_macos(&ca);
             }
-            root_cert = Some(
-                ca_bundle
-                    .into_iter()
-                    .map(|ca| ca.try_into())
-                    .collect::<Result<Vec<_>>>()?,
-            );
+            root_cert = Some(ca_bundle);
         }
 
         match loader.identity(IDENTITY_PASSWORD) {
